feat(serve): implement nursing plan project management with data normalization
- Add normalizeProjectId and normalizeProjectPlans utility functions for data consistency
- Create buildSubmitProjectPlans function to handle proper data conversion for submission
- Normalize project IDs in getProjectList to ensure consistent data format
- Apply data normalization when loading project plans in edit operations
- Convert project plans to proper format before form submission
- Add NursingPlanVo and NursingProjectPlanVo classes for API response structure
- Update NursingPlanService to return NursingPlanVo with project plans included
- Modify controller to return proper R<NursingPlanVo> response type
- Change sortNo field from Long to Integer for proper data typing
- Add NursingProjectPlan domain entity and mapper for plan-project relationships
- Implement batch insert functionality for plan-project associations
- Add SCSS styles for improved UI layout and component styling
- Include proper type conversions between string and numeric project IDs
- Add comprehensive mapper XML configuration for database operations

+++ YiSmartCloud-background/YiSmartCloud-nursing-platform/src/main/java/org/FlyingSparrow/YiSmartCloud/serve/domain/NursingPlan.java
@@ -19,7 +19,7 @@ public class NursingPlan extends BaseEntity {
 
     /** 排序号 */
     @Excel(name = "排序号")
-    private Long sortNo;
+    private Integer sortNo;
 
     /** 名称 */
     @Excel(name = "名称")
